remove exec and fix warnings
